Workaround for datachannel closure
If a detached datachannel does not read from the underlying SCTP stream
it may never receive the message that it's remote has closed the stream.
This causes `Read` to block indefinitely, so we use the deadline and a
boolean to force close the Read call and ensure the stream closes.