swarm: fix DialPeer behaviour for transient connections
This fixes a bug where the first call to `swarm.DialPeer` succeeds and
returns a transient connection with no error while the second call to
DialPeer returns `(nil, network.ErrTransientConn)`.

For dialing, we now only rely on `network.WithForceDirectDial` to force
a direct connection.
For new stream, we open a stream on a transient connection only if
`network.WithUseTransient` is used.